Add response handler
What?
Supports accepting a response handler function in Patova,
this allows us to implement things as training mode based on
Patova's features.

Why?
Sending the rate limit  headers is not the only possible reaction
on error, we might even need to send different headers based on
the limit (for example global vs. local). There are other possible actions
for example not failing the request but limiting the resources put in it
or moving it to a different work-stream it.

One use case is implementing "training mode" where we want to send
the rate limit header but avoid completelly failing the request, that
would let consumers prepare.